+++ Backend/prisma/schema.prisma
@@ -48,8 +48,7 @@ model Produto {
 
 model Pedido {
   Id         String   @id @default(uuid())
-  Mesa       Int
-  Observacao String?
+  Mesa       Int  
   Status     Boolean  @default(false)
   Rascunho   Boolean  @default(true)
   Nome       String?
@@ -64,6 +63,7 @@ model Pedido {
 model ItemPedido {
   Id         String   @id @default(uuid())
   Quantidade Int
+  Observacao String?
   created_at DateTime @default(now())
   updated_at DateTime @default(now())
