+++ Backend/src/routes.ts
@@ -13,8 +13,7 @@ import { ListByCategoryController } from './controllers/product/ListByCategoryCo
 
 import { CreateOrderController } from './controllers/order/CreateOrderController';
 import { RemoveOrderController } from './controllers/order/RemoveOrderController';
-
-import { AddItemController } from './services/category/AddItemController';
+import { AddItemController } from './controllers/order/AddItemController';
 
 import { isAuthenticated } from './middlewares/isAuthenticated';
 
@@ -42,6 +41,6 @@ router.get('/categoria/produto', isAuthenticated, new ListByCategoryController()
 //rotas pedido
 router.post('/pedido', isAuthenticated, new CreateOrderController().handle);
 router.delete('/pedido', isAuthenticated, new RemoveOrderController().handle);
-router.post('/ordemPedido/add'), isAuthenticated, new AddItemController().handle
+router.post('/ordemPedido/add', isAuthenticated, new AddItemController().handle);
 
 export { router }
